Replace size_t with uint32_t

+++ src/allocator.h
@@ -1,8 +1,7 @@
 #pragma once
 #include "multiboot.h"
-#include <stddef.h>
-#include <stdbool.h>
+#include <stdint.h>
 
 #define BLOCK_SIZE 4096
 void allocator_init(multiboot_info_t *multibootInfo);
-void *allocator_allocateBlocks(size_t num);
+void *allocator_allocateBlocks(uint32_t num);
